refactor: split taint flow analysis
